requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
twilio
python-multipart
gunicorn